Remove the PLAYWRIGHT_E2E flag in favour of PHPUnit test selection
