+++ .github/workflows/integration-test.yml
@@ -202,7 +202,6 @@ jobs:
               <php>
                   <env name="APP_ENV" value="test"/>
                   <env name="KERNEL_CLASS" value="App\Kernel"/>
-                  <env name="PLAYWRIGHT_E2E" value="1"/>
                   <env name="PLAYWRIGHT_HEADLESS" value="true"/>
               </php>
           </phpunit>
@@ -395,7 +394,7 @@ jobs:
           echo "Running E2E Tests:"
           echo "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
           echo ""
-          APP_ENV=test PLAYWRIGHT_E2E=1 vendor/bin/phpunit --testdox --colors=always
+          APP_ENV=test vendor/bin/phpunit --testdox --colors=always
           echo ""
           echo "✓ All tests passed"
